+++ src/SadConsole.Shared/Surfaces/BasicSurface.cs
@@ -356,7 +356,7 @@ namespace SadConsole.Surfaces
         /// <param name="file">The destination file.</param>
         public void Save(string file)
         {
-            Serializer.Save((SerializedTypes.BasicSurfaceSerialized)this, file);
+            Serializer.Save((SerializedTypes.BasicSurfaceSerialized)this, file, Settings.SerializationIsCompressed);
         }
 
         /// <summary>
@@ -366,7 +366,7 @@ namespace SadConsole.Surfaces
         /// <returns></returns>
         public static BasicSurface Load(string file)
         {
-            return Serializer.Load<SerializedTypes.BasicSurfaceSerialized>(file);
+            return Serializer.Load<SerializedTypes.BasicSurfaceSerialized>(file, Settings.SerializationIsCompressed);
         }
     }
 }